Changes are imminent
Controller and Filter are now up to date
Controller: PID ignoring width
Filter: weighted average Filter

Controller and Filter are implemented with Commbuffers. Prof. Monetenegro suggested to implement it using method calls, which is smarter.
This is intended as a backup, in case the multithreading-Commbuffer solution is needed at any point in time.

bachelor debug is added to reduce terminal clutter