flow of data was changed+ KF and LQR added
ImageProcessing and UartInterface dont send the data to the filter anymore. The controller recieves it and sends it to the filter.
This was added because the Kalmanfilter needs to know the current controller output. A possible alternative would have been to let the controller return the acceleration after recieve state was called.
KF = Kalman filter. I implemented the extended Kalman filter because our system is not linear. The current KF is not functional
LQR = Linear Quadratic regulator. it doenst work yet but i wanted to push the dataflow changes.